// src/components/tech-atlas/types.ts
export type TechnologyCourseRow = {
    Technology: string;
    Origin_Company_KeyPeople_Place_and_Year: string;
    Simple_Utility: string;
    Real_Life_Example: string;
    Intro_Video_or_Tutorial_Link: string;
    Key_Skills_Thought_Process: string;
    Real_World_Implementation: string;
};

export type DeckViewMode = 'default' | 'sorted' | 'shuffled';